refactor(insight): #451 saju-* → pattern-* rename + pattern-seed-fast-path view 전환
- 파일 rename (git mv): saju-hypothesis(.ts/.test.ts), saju-seed-fast-path.ts, saju-hypothesis-backtest.ts
- 4개 consumer import 경로 갱신 (hypothesis-cards, hypothesis-discovery, weekly-hypothesis-review, insight/index)
- 심볼 rename: trySajuSeedFastPath → tryPatternSeedFastPath
- pattern-seed-fast-path.ts listSeeds SELECT을 pattern_summary view 기반으로 전환
- 스크립트 docstring 실행 명령 갱신

Refs #434
