+++ src/shared/pattern-hypothesis.ts
